broke out bnet example into module
